requirements.txt
asgiref==3.5.2
astroid==2.11.5
black==22.3.0
boto3==1.24.1
botocore==1.27.1
certifi==2022.5.18.1
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.1.3
colorclass==2.2.2
coverage==6.4
cryptography==37.0.2
defusedxml==0.7.1
dill==0.3.5.1
distlib==0.3.4
dj-database-url==0.5.0
Django==3.2
django-active-link==0.1.8
django-admin-thumbnails==0.2.6
django-allauth==0.50.0
django-autoslug==1.9.8
django-countries==7.3.2
django-crispy-forms==1.14.0
django-storages==1.12.3
docopt==0.6.2
filelock==3.7.1
flake8==4.0.1
gunicorn==20.1.0
identify==2.5.1
idna==3.3
isort==5.10.1
jmespath==1.0.0
lazy-object-proxy==1.7.1
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
oauthlib==3.2.0
packaging==21.3
pathspec==0.9.0
Pillow==9.1.1
pip-upgrader==1.4.15
platformdirs==2.5.2
pre-commit==2.19.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.4.0
pylint==2.14.0
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
stripe
terminaltables==3.1.10
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
typing_extensions==4.2.0
urllib3==1.26.9
virtualenv==20.14.1
wrapt==1.14.1
